requirements.txt
flask==2.3.3
flask-cors==4.0.0
deepface==0.0.79
pymongo==4.5.0
pytz==2023.3
Pillow==10.0.0
tf-keras==2.16.0